extract lazy-area.js
