+++ src/index.html
@@ -4,9 +4,6 @@
   </head>
   <body>
 
-    <div id='sample-lazy-load'>
-    </div>
-
     <div id='sample-content'>
     </div>
