--- src/layouts/sample/content-sample.js
@@ -1,5 +1,0 @@
-const debug = require('debug')('spa-starter:content-sample');
-
-debug('content-sample.js has loaded!');
-
-export const loadedDate = new Date();
