+++ src/layouts/sample/index.scss
@@ -2,12 +2,6 @@
 
 /* sample comment */
 
-#sample-lazy-load {
-  border-bottom: 1px dotted black;
-  padding: 1rem;
-}
-
-
 .sample-for-autoprefixer {
   border-radius: 5px;
   display: flex;
