+++ src/layouts/sample/lazy-area.html
@@ -1,0 +1,2 @@
+<div id='sample-lazy-load'>
+</div>
