+++ src/layouts/sample/lazy-area.scss
@@ -1,0 +1,6 @@
+
+#sample-lazy-load {
+  border-top: 1px dotted black;
+  padding: 1rem;
+  margin-top: 1rem;
+}
