+++ src/layouts/sample/lazy-content.js
@@ -1,0 +1,5 @@
+const debug = require('debug')('spa-starter:lazy-content');
+
+debug('lazy-content.js has loaded!');
+
+export const loadedDate = new Date();
